move SVG classes to dedicated package
